backend/requirements.txt
asgiref==3.2.7
astroid==2.3.3
autopep8==1.5
branca==0.3.1
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.41.0
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.1
folium==0.10.1
idna==2.8
inflection==0.4.0
isort==4.3.21
itypes==1.2.0
Jinja2==2.10.3
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.1.2
mccabe==0.6.1
numpy==1.17.3
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.3
pandas==0.25.3
pycodestyle==2.5.0
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.5
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
scipy==1.3.1
seaborn==0.9.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.1
typed-ast==1.4.1
uritemplate==3.0.1
urllib3==1.25.7
wrapt==1.11.2
scikit-learn==0.21.3
scikit-surprise==1.1.0